requirements.txt
black==24.3.0
pandas==2.1.0
pre-commit==3.4.0
pylint==3.1.0
python-dotenv==1.0.0
pyzabbix==1.3.1
requests==2.31.0